much faster explicit agent
